Wrapped the app contents inside TodoApp and Added Input box
